add onlu owners can delete, modify events functionality

+++ src/events/events.controller.ts
@@ -2,6 +2,7 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   HttpCode,
   Logger,
@@ -15,29 +16,19 @@ import {
   UsePipes,
   ValidationPipe,
 } from '@nestjs/common';
-import { CreateEventDto } from './input/create-event-dto';
-import { UpdateEventDto } from './input/update-event.dto';
-import { Event } from './event.entity';
-import { Like, MoreThan, Repository } from 'typeorm';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Attendee } from './attendee.entity';
-import { EventsService } from './events.service';
-import { ListEvents } from './input/list.events';
+import { AuthGuardJwt } from 'src/auth/auth-guard.jwt';
 import { CurrentUser } from 'src/auth/current-user.decorator';
 import { User } from 'src/auth/user.entity';
-import { AuthGuardJwt } from 'src/auth/auth-guard.jwt';
+import { EventsService } from './events.service';
+import { CreateEventDto } from './input/create-event-dto';
+import { ListEvents } from './input/list.events';
+import { UpdateEventDto } from './input/update-event.dto';
 
 @Controller('/events')
 export class EventsController {
   private readonly logger = new Logger(EventsController.name);
 
-  constructor(
-    @InjectRepository(Event)
-    private readonly repository: Repository<Event>,
-    @InjectRepository(Attendee)
-    private readonly attendeeRepository: Repository<Attendee>,
-    private readonly eventsService: EventsService,
-  ) {}
+  constructor(private readonly eventsService: EventsService) {}
 
   @Get()
   @UsePipes(new ValidationPipe({ transform: true }))
@@ -56,47 +47,47 @@ export class EventsController {
 
   // SELECT * FROM event WHERE event.id = 3
   // SELECT * FROM event WHERE event.id > 3
-  @Get('/practice')
-  async practice() {
-    return await this.repository.find({
-      select: ['id', 'when'],
-      where: [
-        {
-          id: MoreThan(3),
-          when: MoreThan(new Date('2021-02-12T13:00:00')),
-        },
-        { description: Like('%meet%') },
-      ],
-      take: 2,
-      order: { id: 'DESC' },
-    });
-  }
+  // @Get('/practice')
+  // async practice() {
+  //   return await this.repository.find({
+  //     select: ['id', 'when'],
+  //     where: [
+  //       {
+  //         id: MoreThan(3),
+  //         when: MoreThan(new Date('2021-02-12T13:00:00')),
+  //       },
+  //       { description: Like('%meet%') },
+  //     ],
+  //     take: 2,
+  //     order: { id: 'DESC' },
+  //   });
+  // }
 
-  @Get('/practice2')
-  async practice2() {
-    // return await this.repository.findOneBy({
-    //   id: 1,
-    // });
+  // @Get('/practice2')
+  // async practice2() {
+  //   // return await this.repository.findOneBy({
+  //   //   id: 1,
+  //   // });
 
-    const event = await this.repository.find({
-      where: { id: 1 },
-      relations: ['attendees'],
-    });
+  //   const event = await this.repository.find({
+  //     where: { id: 1 },
+  //     relations: ['attendees'],
+  //   });
 
-    // const event = new Event();
-    // event.id = 1;
+  //   // const event = new Event();
+  //   // event.id = 1;
 
-    const attendee = new Attendee();
-    attendee.name = 'Using cascade';
-    // attendee.event = event;
+  //   const attendee = new Attendee();
+  //   attendee.name = 'Using cascade';
+  //   // attendee.event = event;
 
-    event[0].attendees.push(attendee);
+  //   event[0].attendees.push(attendee);
 
-    // await this.attendeeRepository.save(attendee);
-    await this.repository.save(event);
+  //   // await this.attendeeRepository.save(attendee);
+  //   await this.repository.save(event);
 
-    return event;
-  }
+  //   return event;
+  // }
 
   @Get(':id')
   async findOne(@Param('id', ParseIntPipe) id: number) {
@@ -117,22 +108,45 @@ export class EventsController {
   }
 
   @Patch(':id')
-  async update(@Param('id') id, @Body() input: UpdateEventDto) {
-    const event = await this.repository.findOne(id);
-    return await this.repository.save({
-      ...event,
-      ...input,
-      when: input.when ? new Date(input.when) : event.when,
-    });
+  @UseGuards(AuthGuardJwt)
+  async update(
+    @Param('id') id,
+    @Body() input: UpdateEventDto,
+    @CurrentUser() user: User,
+  ) {
+    const event = await this.eventsService.getEvent(id);
+
+    if (!event) {
+      throw new NotFoundException();
+    }
+
+    if (event.organizerId !== user.id) {
+      throw new ForbiddenException(
+        null,
+        `You are not authorized to change this event`,
+      );
+    }
+
+    return await this.eventsService.updateEvent(event, input);
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuardJwt)
   @HttpCode(204)
-  async remove(@Param('id') id) {
-    const result = await this.eventsService.deleteEvent(id);
+  async remove(@Param('id') id, @CurrentUser() user: User) {
+    const event = await this.eventsService.getEvent(id);
 
-    if (result?.affected !== 1) {
+    if (!event) {
       throw new NotFoundException();
     }
+
+    if (event.organizerId !== user.id) {
+      throw new ForbiddenException(
+        null,
+        `You are not authorized to remove this event`,
+      );
+    }
+
+    await this.eventsService.deleteEvent(id);
   }
 }
